requirements.txt
Flask==3.0.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
load-dotenv==0.1.0
mysql-connector-python==9.2.0
SQLAlchemy==2.0.37